The hard part comes now

+++ game/server/util.cpp
@@ -1367,7 +1367,7 @@ void UTIL_SetOrigin( CBaseEntity *entity, const Vector &vecOrigin, bool bFireTri
 }
 
 
-void UTIL_ParticleEffect( const Vector &vecOrigin, const Vector &vecDirection, ULONG ulColor, ULONG ulCount )
+void UTIL_ParticleEffect( const Vector &vecOrigin, const Vector &vecDirection, uint32_t ulColor, uint32_t ulCount )
 {
 	Msg( "UTIL_ParticleEffect:  Disabled\n" );
 }
